cyg/doc/XP-130-3-Cube_OEM DocumentationStarterKit_2507/helloIntegrator/requirements.txt
docopt==0.6.2
websocket-client==1.2.1
